Checkpoint commit of day 3 progress.
Main item is AddWord() works with all the latest round of refactorings.
Notable design change was deciding to not store bitmaps or histograms
in the AVL trees at all -- just rely on the 32-bit hashes and have a
means for dealing with hash conflicts.

This trades CPU cost for large space savings.