Implement RemoveWord() and supporting tests.
Plus some tweaks necessary as part of flushing out the RemoveWord()
implementation.  We didn't need the PLENGTH_TABLE_ENTRY pointer in
the WORD_TABLE_ENTRY struct after all, so that frees up a pointer of
space per word entry.

More tests are needed to cover all the behavioral permutations, and
the body of the routine could be cleaned up a bit.